:new:  #1772 电商收付通增加支付结果查询和提现的接口
