+++ weixin-java-pay/src/main/java/com/github/binarywang/wxpay/service/EcommerceService.java
@@ -99,6 +99,18 @@ public interface EcommerceService {
 
   /**
    * <pre>
+   * 合单查询订单API
+   * 文档地址: https://pay.weixin.qq.com/wiki/doc/apiv3/wxpay/pay/combine/chapter3_3.shtml
+   * </pre>
+   *
+   * @param outTradeNo 合单商户订单号
+   * @return 支付订单信息
+   * @throws WxPayException the wx pay exception
+   */
+  CombineTransactionsResult queryCombineTransactions(String outTradeNo) throws WxPayException;
+
+  /**
+   * <pre>
    *  服务商模式普通支付API(APP支付、JSAPI支付、H5支付、NATIVE支付).
    *  请求URL：https://api.mch.weixin.qq.com/v3/pay/partner/transactions/jsapi
    *  文档地址: https://pay.weixin.qq.com/wiki/doc/apiv3/wxpay/pages/transactions_sl.shtml
@@ -138,6 +150,18 @@ public interface EcommerceService {
    * @throws WxPayException the wx pay exception
    */
   PartnerTransactionsNotifyResult parsePartnerNotifyResult(String notifyData, SignatureHeader header) throws WxPayException;
+
+  /**
+   * <pre>
+   * 普通查询订单API
+   * 文档地址: https://pay.weixin.qq.com/wiki/doc/apiv3/wxpay/ecommerce/e_transactions/chapter3_5.shtml
+   * </pre>
+   *
+   * @param request 商户订单信息
+   * @return 支付订单信息
+   * @throws WxPayException the wx pay exception
+   */
+  PartnerTransactionsResult queryPartnerTransactions(PartnerTransactionsQueryRequest request) throws WxPayException;
 
   /**
    * <pre>
@@ -237,4 +261,27 @@ public interface EcommerceService {
    */
   RefundsResult refunds(RefundsRequest request) throws WxPayException;
 
+  /**
+   * <pre>
+   * 二级商户账户余额提现API
+   * 文档地址: https://pay.weixin.qq.com/wiki/doc/apiv3/wxpay/ecommerce/fund/chapter3_2.shtml
+   * </pre>
+   *
+   * @param request 提现请求
+   * @return 返回数据 return withdraw result
+   * @throws WxPayException the wx pay exception
+   */
+  SubWithdrawResult subWithdraw(SubWithdrawRequest request) throws WxPayException;
+
+  /**
+   * <pre>
+   * 电商平台提现API
+   * 文档地址: https://pay.weixin.qq.com/wiki/doc/apiv3/wxpay/ecommerce/fund/chapter3_5.shtml
+   * </pre>
+   *
+   * @param request 提现请求
+   * @return 返回数据 return withdraw result
+   * @throws WxPayException the wx pay exception
+   */
+  SpWithdrawResult spWithdraw(SpWithdrawRequest request) throws WxPayException;
 }
